+++ Plugins/MG/Mammo_CDMam/CDMam_constants.py
