+++ Plugins/MG/Mammo_Flatfield/QCMammo_constants.py
