+++ Plugins/MG/Mammo_Flatfield/QCMammo_lib.py
